Fix Float/Double type mismatch in AnimatedAmount

+++ app/src/main/java/com/example/incometracker/ui/components/CommonComponents.kt
@@ -98,7 +98,7 @@ fun AnimatedAmount(
             targetValue = targetValue,
             animationSpec = tween(durationMillis = 1000, easing = FastOutSlowInEasing)
         ) { value, _ ->
-            animatedValue = value.toDouble()
+            animatedValue = value
         }
     }
